Casting NSNotFound is weird on OS X, too. Fix NSIndexSet unit test.

+++ Tests/NSIndexSetBlocksKitTest.m
@@ -55,7 +55,7 @@
 	};
 	NSUInteger found = [_subject match: indexValidationBlock];
 	STAssertTrue([order isEqualToString: @"123"], @"the index loop order is %@", order);
-	STAssertEquals((NSInteger) found, NSNotFound, @"no items are found");
+	STAssertEquals((NSUInteger)found, (NSUInteger)NSNotFound, @"no items are found");
 }
 
 - (void)testSelect {
